Move primitive types to own crate
