+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
   "cli",
   "client",
+  "core",
   "node",
   "runtime",
   "registry-spec",
