+++ client/Cargo.toml
@@ -10,6 +10,7 @@ license = "GPL-3.0-only"
 repository = "https://github.com/radicle-dev/radicle-registry"
 
 [dependencies]
+radicle-registry-core = { version = "0.0.0", path = "../core" }
 radicle-registry-runtime = { version = "0.0.0", path = "../runtime" }
 
 async-trait = "0.1"
